feat: re-doing `IStream` via `mixins`: part 1
The architectural needs of the library's type were such
that the plain class-based solution no longer did it.
It resulted in code duplication, and poor representation
of the 'Stream' classes. With the new 'mixin' module,
however, one (once again) is in possession of a
more powerful and suitable system for class-formation.

1. refactoring: removing dead code, repetition
2. new exports: "pure" mixins as well as sub-division
	of existing classes onto more new ones
3. simplified inheritance [no more "depth hell" as before...]