cleanup: remove `IStream.prev` operation
It was deemed silly to introduce it, because:

1. The library is (primarily) intended for creation of Predictive parsers,
and not Backtracking ones.
2. The backtracking operation functions does not coexist with the
chosen `RecursiveList` algorithm, due to the necessity of `IStream`s
to return a value. An `IStream` that would merely be verifying a path
for backtracking cannot return a value to the underlying Stream. Even if
one were to enable such streams, the implications for the logical complexity
of parsers would be horrendous - the benefit of simple, uni-directional
functional would be (at large) lost. There would be no point in even using
the library.

List of changes:

1. removed `prev` method (from everywhere except `DepthStream`)
2. removed `rewind` method (from everywhere except `InputStream` and `DepthStream`)
3. removed the `FreezableStream` class
4. removed the `isCurrStart` method (from everywhere except `DepthStream`)
5. removed `ILineIndex.prevChar` method, and its implementations
6. renamed `BackupIndex` into `StringLineIndex`
7. removed `RotationBuffer.prototype.backward` method, and all related ones